+++ module/core/telemetry/config.go
@@ -6,7 +6,7 @@ const (
 	// defaultServiceName defines fallback telemetry service-name values.
 	defaultServiceName = "mannaiah-api"
 	// defaultServiceVersion defines fallback telemetry service-version values.
-	defaultServiceVersion = "v2.0.8"
+	defaultServiceVersion = "v2.0.9"
 	// defaultMetricsPath defines fallback Prometheus metrics endpoint paths.
 	defaultMetricsPath = "/metrics"
 	// defaultTracesExporter defines fallback traces exporter values.